refactor components in preparation for reminders work
